requirements.txt
kafka-python==2.0.2
pymongo==4.2.0
confluent_kafka==2.3.0